Added User Signup
Added ability for signing up. Added a few other files to be modified later.

+++ api/routes/users.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const router = express.Router();
 const mongoose = require("mongoose");
+const bcrypt = require("bcrypt");
 
 const User = require("../models/user");
 
@@ -18,9 +19,10 @@ router.get("/", (req, res, next) => {
             email: doc.email,
             role: doc.role,
             _id: doc._id,
+            userImage: doc.userImage,
             request: {
               type: "GET",
-              url: "http://localhost:3000/users/" + doc._id,
+              url: "http://localhost:3000/api/users/" + doc._id,
             },
           };
         }),
@@ -33,41 +35,54 @@ router.get("/", (req, res, next) => {
     });
 });
 
-router.post("/", (req, res, next) => {
-  const newUser = new User({
-    _id: new mongoose.Types.ObjectId(),
-    firstName: req.body.firstName,
-    lastName: req.body.lastName,
-    email: req.body.email,
-    password: req.body.password,
-    role: req.body.role,
-  });
-
-  newUser
-    .save()
-    .then((result) => {
-      console.log(result);
-      res.status(201).json({
-        createdUser: {
-          firstName: result.firstName,
-          lastName: result.lastName,
-          email: result.email,
-          role: result.role,
-          _id: result._id,
-          request: {
-            type: "GET",
-            url: "http://localhost:3000/users/" + result._id,
-          },
-        },
-      });
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json({ error: err });
+router.post("/signup", (req, res, next) => {
+  User.findOne({ email: req.body.email })
+    .exec()
+    .then((user) => {
+      if (user) {
+        console.log(user);
+        return res.status(409).json({
+          message: "User already exists",
+        });
+      } else {
+        bcrypt.hash(req.body.password, 10, (err, hash) => {
+          if (err) {
+            console.log(err);
+            return res.status(500).json({ error: err });
+          } else {
+            const newUser = new User({
+              _id: new mongoose.Types.ObjectId(),
+              firstName: req.body.firstName,
+              lastName: req.body.lastName,
+              // userImage: req.file.path,
+              email: req.body.email,
+              password: hash,
+              role: req.body.role,
+            });
+            newUser
+              .save()
+              .then((result) => {
+                console.log(result);
+                res.status(201).json({
+                  message: "User created",
+                  request: {
+                    type: "GET",
+                    url: "http://localhost:3000/api/users/" + result._id,
+                  },
+                });
+              })
+              .catch((err) => {
+                console.log(err);
+                res.status(500).json({ error: err });
+              });
+          }
+        });
+      }
     });
 });
 
 router.get("/:userId", (req, res, next) => {
+  console.log(req.params.userId);
   const id = req.params.userId;
   User.findById(id)
     .select("firstName lastName email role")
@@ -80,8 +95,11 @@ router.get("/:userId", (req, res, next) => {
           lastName: doc.lastName,
           email: doc.email,
           role: doc.role,
+          _id: doc._id,
+          userImage: doc.userImage,
         });
       } else {
+        console.log("No valid entry found for provided ID");
         res
           .status(404)
           .json({ message: "No valid entry found for provided ID" });
@@ -106,7 +124,7 @@ router.patch("/:userId", (req, res, next) => {
         message: "User updated",
         request: {
           type: "GET",
-          url: "http://localhost:3000/users/" + id,
+          url: "http://localhost:3000/api/users/" + id,
         },
       });
     })
@@ -118,16 +136,26 @@ router.patch("/:userId", (req, res, next) => {
 
 router.delete("/:userId", (req, res, next) => {
   const id = req.params.userId;
-  User.deleteOne({ _id: id })
+  User.findById(id)
     .exec()
-    .then((result) => {
-      res.status(200).json({
-        message: "User deleted",
-      });
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json({ error: err });
+    .then((user) => {
+      if (!user) {
+        return res.status(404).json({
+          message: "User not found",
+        });
+      } else {
+        User.deleteOne({ _id: id })
+          .exec()
+          .then((result) => {
+            res.status(200).json({
+              message: "User deleted",
+            });
+          })
+          .catch((err) => {
+            console.log(err);
+            res.status(500).json({ error: err });
+          });
+      }
     });
 });
